+++ src/test/java/lotto/model/dto/WinningResultTest.java
@@ -1,7 +1,8 @@
-package lotto.model;
+package lotto.model.dto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import lotto.model.WinningRule;
 import org.junit.jupiter.api.Test;
 
 public class WinningResultTest {
